+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 
 ### Added
 - Bars can now be built by `beginTime` instead of `endTime`
+- Added `NumericIndicator#openPrice`
+- enhanced `ROCIndicator` with  `previousIndicator` to set the type of the previous indicator
 
 
 ## 0.18 (released May 15, 2025)
@@ -106,7 +108,6 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - Added `KiJunV2Indicator` to `indicators.averages`-package: Kihon Moving Average (KiJunV2) indicator
 - Added `VIDYAIndicator` to `indicators.averages`-package: Chande’s Variable Index Dynamic Moving Average (VIDYA) indicator
 - Added `VWMAIndicator` to `indicators.averages`-package: Volume Weighted Moving Average (VWMA) indicator
-- added `AverageIndicator`
 
 ## 0.17 (released September 9, 2024)